Extracts PrimaryCalendarSelector to DestinationCalendarSelector
